refactor(pages): 탭 컴포넌트 구조 개선 및 Suspense 적용
